net: axienet: Fix reading of "xlnx,addrwidth" DT property
"xlnx,addrwidth" DT property is of 32-bit, using of_property_read_u32
to read the value.

Message-ID: <20250321130257.4045259-1-ravikanth.tuniki@amd.com>
State: pending

+++ drivers/net/ethernet/xilinx/xilinx_axienet.h
@@ -1118,7 +1118,7 @@ struct axienet_local {
 	u16 chan_id;
 	u16 weight;
 
-	u8 dma_mask;
+	u32 dma_mask;
 	u32 usxgmii_rate;
 
 	u32 max_speed;		/* Max MAC speed */
